+++ src/sortCompact/index.ts
@@ -1,0 +1,8 @@
+/**
+ * sortCompact sort the given array using native JS sort function, and
+ * in addition it also compacts the resulted array. It removes all falsey values except the
+ * number 0.
+ */
+const sortCompact = (array: Array<any>, sortFunction?: (left: any, right: any) => boolean): Array<any> => {};
+
+export default sortCompact;