+++ utils/prettyTestName.ts
@@ -1,0 +1,14 @@
+export default (text: string, received: any, expected: any, test: ) => {
+  let stringyfiedReceived = received;
+  let stringyfiedExpected = expected;
+  try {
+    stringyfiedReceived = JSON.stringify(stringyfiedReceived);
+  }
+  try {
+    stringyfiedExpected = JSON.stringify(stringyfiedExpected);
+  }
+
+  const receivedText = text.replace("%s", stringyfiedReceived);
+  const expectedText = receivedText.replace("%s", stringyfiedExpected);
+  return expectedText || receivedText;
+}